// React/app/GerAdv_TS/AlarmSMS/Interfaces/interface.AlarmSMS.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IAlarmSMS {
// 202501251
    id: number;
 
	operador: number,
	agenda: number,
	recado: number,
	descricao: string,
	hora: number,
	minuto: number,
	d1: boolean,
	d2: boolean,
	d3: boolean,
	d4: boolean,
	d5: boolean,
	d6: boolean,
	d7: boolean,
	email: string,
	desativar: boolean,
	today: string,
	excetodiasfelizes: boolean,
	desktop: boolean,
	alertardatahora: string,
	guidexo: string,
	emocao: number,
}

export interface IAlarmSMSFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IAlarmSMSIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}